fixing get_all_auctions_menu
